Suggested Fee Changes, not checked very well
,
		{
			height: 1,626,337,
			fees: {
				send: 330000,		          // 0.0033
				vote: 3300000,	                 // 0.033
				secondsignature: 10000000,	// 0.1
				delegate: 6000000000,	       // 60
				multisignature: 50000000,     // 0.5
				dapp: 2500000000	     //25
			}
		}
	],


Send Fee should be reduced to increase the number of transactions. I have suggested a 10x decrease to 0.0033 or about 1 cent
The Voting fee has been a bit problem, here I have decreased it to 0.033 (about 10 cents) which mirrors the number of delegates a single vote is worth. Voting in and of itself benefits the community. The issue this could lead to is pool hopping but that should be solved on payout side in my opinion.
The Secondsignature increases account security and it seems like most are willing to pay it. 
I really think the Delegate Fee should Remain the same or even increase. Here I have left them the same.
I left the Multisignature and dapp Fees the same and the second signature. I have not heard any complaints about them. 
Currently the dapp fee is a little high.